pyproject.toml
[project]
name = "agentic"
version = "0.1.0"
description = "A simple AI coding assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
